Move NdIter to its own file
