+++ candle-core/src/lib.rs
@@ -71,6 +71,7 @@ pub mod layout;
 pub mod metal_backend;
 #[cfg(feature = "mkl")]
 mod mkl;
+pub mod nditer;
 pub mod npy;
 pub mod op;
 pub mod pickle;
@@ -101,6 +102,7 @@ pub use dummy_dtype::{F4, F6E2M3, F6E3M2, F8E8M0};
 pub use error::{Context, Error, Result};
 pub use indexer::{IndexOp, TensorIndexer};
 pub use layout::Layout;
+pub use nditer::NdIter;
 pub use shape::{Shape, D};
 pub use storage::Storage;
 pub use streaming::{StreamTensor, StreamingBinOp, StreamingModule};
